This is the second patch in a series intended to improve the handling of IO errors. In this patch I have added a parsing of the ildg-format LIME record. This record is now parsed and checked versus the required input parameters. The following 4 error cases are (more properly) taken care of in this patch:
1) Gauge file does not contain ildg-format record/Lime record "ildg-format" is malformed in gauge file
This causes an error exit. This behaviour is new.

2) Multiple ildg-format records in gauge file
This causes an error exit. This behaviour is new.

3) LIME record ildg-format does not match hmc.input for precision or dimensions
This is a rather perverse error case, as in general the problem should already be apparent
at the reading of the binary part (which happens first). However, there are subtle
border cases such as a 4^4 lattice in double precision being read as a 4^3x8 lattice
in single precision that can be caught here, as well as a mismatch between the ildg-format
and the ildg-binary-data record. This behaviour is new.

4) LIME record ildg-binary-data does not match hmc.input for precision and dimensions
This causes an error exit, it used to do so already, but now the message is more descriptive.
This behaviour is not new, although there are some borderline case as described in 3) where
the code used to override user input and dynamically set precision in order to proceed.
That is now no longer allowed.

The parsing of the ildg-format record allows the clean separation of read parameters and 
input parameters. I have tried to include this in the gauge reading code in a clear way.
This means that the gauge reading now takes an extra parameter, both in the Lime and Lemon case
and does not rely on the availability of global variables as before.

Precision for reading in a gauge configuration is an input parameter, but also used to
be computed dynamically inside the gauge reading code. In this patch, the dynamic part is gone.

Some minor cosmetic error reporting changes throughout, such as indenting with #.

The paramsIldgFormat struct used nx,ny,nz and nt to describe global lattice dimensions before.
This might be confused with NrXProcs etc, so I have renamed it to lx,ly,lz and lt. This renaming
forced minor changes in other parts of the code, such as io/params.h and io/gauge_write_format.c

Minor typo corrected in doc/input.tex (double double)

In this patched version, the code now refuses to run when given any of the configurations
identified as broken by Andreas Nube. I believe that is the intended behaviour.

Please compile and test.

Still planned:
- Propagator checks
- Readback checks of gauge/propagator writing
- Fallback behaviour Lemon -> in case of failed readback checks
- Unified error code system (needed to properly implement fallback behaviour)