This is the third patch in a series intended to improve the handling of IO errors.
This patch focuses on the reporting part of errors, though other changes were also
made. There are a lot of textual changes and whitespace fixes in this patch, which
makes it look more scary then it is in reality. The motivation for these is that
when looking at debuglevel = 0 at the output of the code, it is a mess.
Capitalization, start of line characters, exclamation points were all used without
any visible organisation. Some output which does not appear until debuglevel = 1
is really needed to understand what is reported at debuglevel = 0.

Initially I intended to look only at invert, but the large variety of operators,
spinor writers and others made that an impossible task. Therefore, this patch is
still incomplete.

Textual changes aside, there are some real improvements in this patch, such as
real error checking on the writing part of the code. More details below.

- Unified capitalization of even/odd in invert_doublet_eo.c, invert_overlap.c,
  invert_eo.c and io/utils_write_first_message.c
- Changed the default input filename for invert to invert.input. Looking at the
  sample files and the usage that I am aware of, everybody seems to be using this
  and compared to the name for the hmc default, it seems sensible. This change was
  also made to doc/input.tex
- Changed io/utils_write_first_message.c to active form where passive was used. In
  the case of compilation, this looked ok, but for the preconditioning, it makes
  little sense to used the past tense, when the code has not even begun doing
  anything.
- MANY whitespace corrections in operator.c
- In operator.c, the timing for the writing of the propagator was removed, this
  should be handled by the writing code such as io/spinor_write_binary.c. The
  timing of the inversion itself was left in of course.
- Many whitespace changes in invert.c
- Some parts of invert.c give an error exit with error code 0, this I changed to
  exit(-1) as exit(0) generally means success.
- I now explicitly say that the plaquette value of a read gauge field is _computed_,
  in anticipation of a future patch where the computed value will be compared with
  a value read from LIME headers.
- An unused variable (modenumber) was removed. Feel free to add it back in when
  needed.
- Many whitespace changes in hmc_tm.c
- When writing a gauge field produces an error, that error is now caught, and an
  error exit is produced. This is new behaviour, and a place where a future
  fallback routine should be implemented. A change in io/gauge_write.c was needed
  to make this possible.
- If moving conf.tmp to conf.save/conf.#### fails, that is error is now caught
  and an error exit is produced. This is new.
- In spinor_write_binary.c if a C-LIME call tries to write a spinor and produces
  and error, that error is now caught (new) and passed on to the calling
  function (new).

A further note: aggressive error checking based on scidac-checksums, as aggressive
as done for gauge files is currently not entirely possible for propagators and
sources, based on the way we currently structure our lime files. The combination
of a scidac-checksum and scidac-binary-data record is currently a bit ambiguous 
(they are directly following eachother). I will try to move those together into a
single LIME message (as multiple records) in a later patch.

Finally, testing error cases for reading is easier than testing for writeout errors.
Some of the new code has only seen rudimentary testing (manual insertion of error
codes).