This is the sixth patch in a series intended to clean up the handling of IO errors.
Following a discussion amongst Albert, Carsten and me, we have decided to leave
the implementation of a fallback mode for Lemon IO to Lime IO. This would add
substantial complications in the code for limited benefit over the chosen, and
in this patch implemented solution.

- hmc_tm now verifies a Lemon written file by reading it in again. If any
  problems are found, the code exits with a warning and error code -3. The
  broken configuration will be left on disk with name .conf.tmp.
- Writeout of configurations is now better described in the main code
- SciDAC checksum now are always properly formatted, with leading zeroes
- Both SciDAC checksum and ILDG format LIME records are now verified at all
  debug levels, the following input parameter (not implemented in this patch)
  will govern whether they are active or not:
  DisableChecks = No (default) / Yes

Please note that this patch implements a crucial step in the fix for the
observed problems with gauge IO. With this patch applied we can once again
guarantee integrity of the gauge files on disk. While we have not solved or
identified the underlying cause, this should make it safe to turn Lemon back
on again in production runs.