added the input value
SourceTimeSlice = detect

in order to allow automatical detection of the timeslice value in the
inversion. This value is used to construct the filename to read the
source from. If you want to run more than one gauge in a single
inversion, this feature is needed. An example input file would be
(only the relevant input parameters)

Measurements = 20
Nsave = 8
InitialStoreCounter = 1020
SourceTimeSlice = detect
ReadSource = yes

would do 20 gauges, starting with conf.1020 in steps of 8. And it
expects sources as files like

source.0742.41.00

and would detect 41 as timeslice.