first running version of cloverdetratio monomial, no test yet but HMC gives acceptance. The input file has the following new monomial structure
BeginMonomial CLOVERDETRATIO
  Timescale = 1
  2KappaMu = 0.50
  rho = 0.5
  rho2 = 0.1
  CSW = 1.00
  kappa = 0.138
  AcceptancePrecision =  1.e-20
  ForcePrecision = 1.e-12
  Name = cloverdetratio
  solver = CG
EndMonomial

where rho is the preconditioning shift in the nominator and rho2 is
the shift in the denominator (see docu). Also the CLOVERDET monomial
has a new parameter rho for the preconditioning shift, e.g.

BeginMonomial CLOVERDET
  Timescale = 1
  2KappaMu = 0.50
  rho = 0.5
  CSW = 1.00
  kappa = 0.138
  AcceptancePrecision =  1.e-20
  ForcePrecision = 1.e-12
  Name = cloverdet
  solver = CG
EndMonomial

This implementation can be improved by making use of the fact that
delta hat Q and delta hat W are identical and independent of rho and
rho2, as explained in the docu. Currently this is the same
implementation as used for the DETRATIO monomial.