bug for XYT parallel fixed, we need more gauge fields.
This needs better treatment, as currently the parameter
g_dbw2rand
indicates whether extra boundaries are exchanged or not. I had to
change the input file such that it does not set g_dbw2rand =0
anywhere, because that might lead to problems as the order of
evaluation is not fixed.