this fixes the problem with the online CORRELATOR measurement. Now one needs to specify an operator in the input file for which the inversions will be performed.
Currently, only the first operator in the file will be used and only
TMWILSON, WILSON and CLOVER are working. For other operators the
measurement will not be done and the hmc proceeds as if no CORRELATOR measurement
specified in the input file.

Also, removed MAX_ITER_ variables from global.h and all other appearences, as it
was not longer needed. Therefore, also the input parameters

BCGstabMaxIter =
CGMaxIter =

are no longer supported and removed from input.