Change the indices $F^{\alpha\beta}$ → $F^{ab}$
The gluon field strength tensor should have a color indices in the
adjoint representation. It also has Lorentz indices. The α and β read
like Dirac indices which I think should not occur on the gluon field
strength tensor.

The Lorentz indices μ and ν are summed over, the left side of the
equation has a, α, b, and β as free indices. Therefore there need to be
some indices a and b on the right side.