Modified reorder_gauge_toQhix to unable gauge compression.
Now, we are only copying 2 rows to the QPhiX gauge field when gauge
compression is enable for the dslash. Note that the shorter color dimension
is the slower one in QPhiX and the faster one in tmlQCD (since they are
transposed w.r.t. each other).

In QPhiX, one gauge field element is the Structure of Arrays:
SU3MatrixBlock[8][ ( compressP ? 2 : 3 ) ][3][2][V]