Add Travis CI support
To actually build on Travis CI, go to https://travis-ci.org/etmc/tmLQCD
and log in with an account having admin rights on that repository. There
should be a switch to enable building. Then it is automatically
integrated with GitHub.

Currently it uses one of the HMC input files to run it. It would be even
better to use some proper unit test. The current usage is a “smoke test”
(as in “Does it smoke when it is turned on?”) only.