qphix_interface: Remove old full_clover_product
The `full_clover_product` function assumed a few things about the layout
which were wrong. There is a new implementation in `clover_product`
which can handle both `CloverBlock` and `FullCloverBlock`.

The twisted mass clover multiplication will be fixed by using the new
function in the virtual functions. Now it just calls `clover_product`
and pass it a `FourSpinorBlock *` and the template argument deduction
will perform the actual product with the appropriate specialization of
`InnerCloverProduct::multiply`.