This is a fix for issue #393
The problem are the explicit SSE version of non-even/odd D_psi and
Dsw_psi.
The only feasible workaround at the moment is to disable SSE2|3 in
D_psi.
This will not affect the even/odd versions, which are still available
in SSE2|3.

In the future we'll remove these macros completely.